refactor(ui): extract context menu item builder
